fix (display issue): fixed a part of the display on windows.

+++ src/server/systems/Systems.cpp
@@ -386,7 +386,7 @@ void Systems::collision_system(Registry& reg,
 }
 
 void Systems::shoot_system(Registry& reg,
-                           u_int32_t clientId,
+                           size_t clientId,
                            entity_t playerId,
                            std::unique_ptr<NetworkSender>& networkSender,
                            RType::Logger& logger) {
